+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Features
 
+- [#3065](https://github.com/poanetwork/blockscout/pull/3065) - Transactions history chart
+
 ### Fixes
 
 - [#3064](https://github.com/poanetwork/blockscout/pull/3064) - Automatically define Block reward contract address in TokenBridge supply module